Rename some supporting files to match content. Move some things around.
