--- actors/builtin/miner/partitions.go
@@ -1,7 +1,0 @@
-package miner
-
-// Identifier for a single partition within a miner.
-type PartitionKey struct {
-	Deadline uint64
-	Partition uint64
-}
